+++ packages/framework-integration-tests/integration/fixtures/common/cart-item-with-fields.ts
