+++ packages/framework-integration-tests/integration/fixtures/entities/post-with-fields.ts
