+++ packages/framework-integration-tests/integration/fixtures/entities/post-with-reducer.ts
@@ -1,18 +1,13 @@
 import { Entity, Reduces } from '@boostercloud/framework-core'
 import { UUID } from '@boostercloud/framework-types'
-import { PostCreated } from '../events/PostCreated'
+import { PostCreated } from '../events/post-created'
 
 @Entity
 export class PostWithReducer {
-  public constructor(
-    public id: UUID,
-    readonly title: string,
-    readonly body: string,
-  ) {}
+  public constructor(public id: UUID, readonly title: string, readonly body: string) {}
 
   @Reduces(PostCreated)
   public static reducePostCreated(event: PostCreated, currentPostWithReducer?: PostWithReducer): PostWithReducer {
     return /* NEW PostWithReducer HERE */
   }
-
 }
