+++ packages/framework-integration-tests/integration/fixtures/events/cart-changed-with-fields.ts
