+++ packages/framework-integration-tests/integration/fixtures/read-models/cart-read-model.ts
