+++ packages/framework-integration-tests/integration/fixtures/read-models/cart-with-fields-read-model.ts
