+++ packages/framework-integration-tests/integration/fixtures/read-models/cart-with-projection-read-model.ts
@@ -1,6 +1,6 @@
 import { ReadModel, Projects } from '@boostercloud/framework-core'
 import { UUID, ProjectionResult } from '@boostercloud/framework-types'
-import { Cart } from '../entities/Cart'
+import { Cart } from '../entities/cart'
 
 @ReadModel({
   authorize: // Specify authorized roles here. Use 'all' to authorize anyone
